requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py38h578d9bd_0
charset-normalizer=2.1.0=pypi_0
cudatoolkit=11.1.1=ha002fc5_10
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.0=pypi_0
freetype=2.10.4=h0708190_1
gdown=4.5.1=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.10.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.47.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py38h06a4308_0
ipython=8.4.0=py38h06a4308_0
jedi=0.18.1=py38h06a4308_1
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jupyter_client=7.2.2=py38h06a4308_0
jupyter_core=4.10.0=py38h06a4308_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.43.0=h7f98852_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py38h06a4308_0
nettle=3.6=he412f7d_0
ninja=1.11.0=h924138e_0
numpy=1.19.4=pypi_0
numpy-base=1.23.1=py38ha15fc14_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openssl=1.1.1o=h166bdaf_0
pandas=1.4.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.1.2=py38hb39fc2d_0
pip=22.1.2=py38h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pysocks=1.7.1=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2022.1=pypi_0
pyzmq=23.2.0=py38h6a678d5_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.0=pypi_0
setuptools=61.2.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.2=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu111
tornado=6.1=py38h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.11=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=ha95c52a_0
